+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Library API: `sll_library_handle_t`, `sll_platform_load_library`, `sll_platform_lookup_function`, `sll_platform_unload_library` and `sll_api_sys_load_library` ([#123])
 - Logging API: `SLL_LOG`, `sll_log`, `sll_set_log_default`, `sll_set_log_file` and `sll_set_log_function` ([#126], [#128] and [#129])
 - Max internal function length: `SLL_INTERNAL_FUNCTION_MAX_LENGTH`
+- Sll logging API: `sll_api_log_log`, `sll_api_log_set_default`, `sll_api_log_set_file`, `sll_api_log_set_func`, `sll:log_log`, `sll:log_set_default`, `sll:log_set_file` and `sll:log_set_func` ([#126] and [#127])
 - Sll version API: `sll_api_sys_get_version` and `sll:sys_get_version`
 - Website code in [`src/web`][main/src/web]
 
@@ -1168,6 +1169,7 @@ Unfortunately, no versions were assigned before 2021-6-15 (:disappointed:), so t
 [0.1.2]: https://github.com/sl-lang/sll/tree/v0.1.2
 [#129]: https://github.com/sl-lang/sll/issues/129
 [#128]: https://github.com/sl-lang/sll/issues/128
+[#127]: https://github.com/sl-lang/sll/issues/127
 [#126]: https://github.com/sl-lang/sll/issues/126
 [#125]: https://github.com/sl-lang/sll/issues/125
 [#124]: https://github.com/sl-lang/sll/issues/124